Partially implement transfering of edge data from lights to shadows
This is an optimization. The idea is that the lights will figure out
exactly what line segments are in view, and only create shadows for
those line segments. (Previously, it created shadows for all shapes,
which was too slow since one of the algorithms is O(n^2), where n is the
number of shapes).