Fix bug in RotatableLight
The problem was that if the front face of a shadow is very short, it
culls the front face as well as the side faces, without checking how
long the side faces are. In some edge cases, the side faces are long
enough to be noticeable when they are culled, even though the front face
is very short.