Refactor snapping system
Remove the SnappableObject class (which requires you inherit from it).
Add the Snapper class, which is its own component which performs the
snapping independently of whatever else is in the gameobject. The
MonoBehaviours with more complex snapping will now find the Snapper
component on its gameobject and use that to do its special snapping.